Fix ALT-TAB and keyboard activated menu issues
Fixes repeating keys in OS9 when entering menus using keyboard
and other anomolies related to use of Vcc in a windowed environment.

ALT keys that are not configured to access a Vcc menu can be used in
OS-9 to generate special graphics characters and as a control in some
OS-9 applications, for example full screen text editors.

It is possible to turn of Vcc menu access vi ALT keys off.

Currently left ALT menu access is hardcoded off and RIGHT ALT on.
This should allow Vcc running in windowed mode to behave as well as
full screen mode for ALT key use in OS-9 and NitrOS-9 yet still
provide keyboard activated Vcc menu access via the right ALT key.

A future enhancement will be to add config controls to allow users
to modify ALT key function to suit their desires.

All keys, including both ALT keys, are sent to the emulation.
In RSBASIC either ALT key will generate an '@' on the screen even
when used to access a Vcc menu.

This change does not prevent ALT-TAB window switching but does
fix OS-9 keyboard issues on return.

This commit is a further enhancement.