Zero fill if read past end of VHD file
A check was previously added to return an error to the CoCo OS if an
attempt was made to read past the end of a VHD file.  This broke the
ability to format an empty VHD file on [Nitr]OS9. This change returns
data containing zeros instead of an error after EOF is reached.