PasteDelay Adjustment
This field sets the interkey delay - the delay in seconds between KEYDOWN and KEYUP.  It has to be long enough for BASIC to recognize the key press but short enough so that the paste doesn't take forever.  This is greatly influenced by the frame rate.  In debug was have a lower frame rate so we need a slightly higher value.  In release we need a lower value.