Fixed Clipboard Paste Feature (#119)
* Fix for Access Exception on Cleanup

We were not waiting for the Emulation Thread to finish up so Chet's critical section could be in use when the thread was terminated.  This was made more likely with the Execution Trace using it every CPU Cycle.

* Fixed the Clipboard Paste feature and reduced the Frame Loading by ExecutionTrace

* Changed out the CriticalSections to std::atomic<bool> for certain bools used by the CPUs - this provided a good frame boost.
* Fixed the clipboard paste to use an independent paste timer for more reliable pasting.
* Cleaned up some Execution Trace logic to handle conditions when the trace collects samples so fast that the Execution Trace Window doesn't realize it has completed.

* PasteDelay Adjustment

This field sets the interkey delay - the delay in seconds between KEYDOWN and KEYUP.  It has to be long enough for BASIC to recognize the key press but short enough so that the paste doesn't take forever.  This is greatly influenced by the frame rate.  In debug was have a lower frame rate so we need a slightly higher value.  In release we need a lower value.

* Cut and Paste error