More work on acia file i/o. Cleaner thread exits.
Output thread sometimes exiting before all data is written because
TerminateThread was being used to destroy threads when DTR was
dropped.  Re-worked to use thread stop events to notify threads
when it is time for them to exit. Output thread can then flush
the output buffer before exiting. TerminateThread is used only
if an I/O thread fails to exit on it's own. (On XP TerminateThread
can cause resource leaks)