Convert User Guide to markdown
User guide converted "Welcome to VCC <vers>.pdf" to UserGuide.md and
placed into git. Images for the document are stored in docs/images.
The guide had been maintained as a windows document and manually
converted to a pdf to be included in release zip files.
This change allows the document to be saved in git as text so changes
can be tracked.  Markdown will be converted to .html and then to
pdf manually for the release files.