Allow physical addressing in disassembly window (#180)
* Initial commit physical mem disassembler

This change allows disassembly of all of physical memory relative
to a given block number. Any reads beyond physical memory will
return all bits set, ($FF or STU $FFFF).   This allows disassembly
of code that is not currently in CPU memory.

This portion adds ability to disassemble either CPU or physical memory
based on a flag and block number in the debugger CPUTrace structure.

Still to do is add check box for physical mem and box to enter
the block number.

* Disassemble physical memory working

Might need a tweak to improve interface.

* Fix a typo

* Put physical address in disassembly window

When physical addressing is selected the address in the disassembly
lines are the physical address instead of the offset from start of
the block.

Also added tabstop formatting to output edit box to better control
alignment of disassembly fields.

* Disassembly put count lines instead of to address

The to address edit box is replaced with a line count, Disassembly
continues until count instruction lines are disassembled.

Improve tab between edit boxes.

---------

Co-authored-by: ejaquay <ejjaquay@gmail.com>