GUI changes to permit setting breakpoints
Single clicking or move up/down in decoded text sets current line
position. Once line is set pressing 'B' or 'b' will be used to set
breakpoint on that line.  Breakpoint address color is set to red.
If 'C' or 'c' is pressed breakpoint will be cleared and color
is reset.

Still to do is saving breakpoints in an array and implementing them.