Automatic removal of haltpoints on halt
When Vcc is halted any haltpoints that were placed (opcode in memory
replaced by 0x15) are removed and the original opcodes are restored.
This is not yet bullet proof - additional checks for resets may be
required

Halts and Breaks are now considered seperate events and are treated
differently.  Halts are placed by the Vcc disassembler screen while
Breaks are code added during assembly using lwasm.  Halts are removed
when executed, breaks are not.

Hotkey changes:
  H To add haltpoint
  R To remove haltpoint
  S to print haltpoints and their status
  P to place haltpoints into memory
  U to remove haltpoints from memory

Added checks to prevent placing a haltpoint when one is already
there and to prevent removing any that were not.