Instruction stepping working after a halt.
This required changes to hd6309.c and mc6809.c to allow a single
instruction to execute before re-applying haltpoints. Hopefully
nothing got broken in the process.  In mc6809.c execute a single
instruction placed into a seperate function. Someday these
should be changed to defined jump vectors instead of relying
on the compiler to do it (like hc6309 does).