Gime timer enhanments to improve game sounds.
Many games require the gime to set the timer interupt flag(s) even if
the timer is set to zero.  This change foces the timer setting to be
at least one so the interupt occurs even if a zero timer value is set.
Many games require this to generate sounds.