Prevent Memory Display access exceptions.
Using CPU config to change Memory Size or using F9 to reset VCC
while the Memory Display is running was found to cause occasional
access violations. MMU init code was relocating memory at the same
time SafeMemRead8 was reading it.

This fix adds an atomic initializing flag to MmuInit() and
SafeMemRead8() to greatly reduce the possibility of conflict.