SDC resolved FD502 contention.
The code that switched floppy ports in mpi.cpp to the spare slot
select was commented out long in the past for reasons unknown
(before github).

Added #define BANKED_CART_OFFSET to re-enable the commented code
and the floppy contention was resolved.  If the spare slot select
for floppy ports proves to be a problem a work-around will be
required for cart dll's to request this from MPI.