Clean up COCO3.ROM load code
Several non-working code fragments that were used in the past to load
alternate versions of COCO3.ROM are removed.  COCO3.ROM is expected in
the directory VCC.exe is in.  Alternately the full path to a different
rom can be specified in the vcc.ini file in the "DefaultPaths" section
with the tag "COCO3ROMPath"

If the alternate rom is specified and it fails to load or if is not
specified and COCO3.ROM in the VCC.exe directory fails to load then
Vcc.exe fails to start with a missing coco3 rom messagebox.