Modify acia to compile as C++ (#372)
* Change SDC dll dialog unload order.

Close dialogs before closing drives.

* Change extension of DialogOps and compile attributes for acia

For now the acia module is still compiled using C rather than C++.
DialogOps was created with a .c extension so as to not confuse the C
compiler when building acia.  This commit changes DialogOps.c to
DialogOps.cpp and uses CompileAsC attribute in acia.vcxproj instead.

* Rename .C files to .CPP

Rename files befing compiled as C++ from XXX.c to XXX.cpp to make
which compiler is buing used more obvious.

* Converting acia to C++

in progress ...

* Modify acia code to compile with C++ fixes

Fixes from testing modified acia code.

Note: These is still a problem with the AssertInterupt.  This will
be addressed on another commit as it involves changes to base code.

---------

Co-authored-by: ejaquay <ejjaquay@gmail.com>