Feature: Export memory region to disk
Replaces the single address box with an address range, two addresses, one for the beginning of the range and one for the end of the range.

Add a button to export the entered range to disk. Pop up a file dialog for selection of the save filename.

Remove the Find button, just rely on pressing enter/tab within the address range fields.