Add reset button to mpi config screen
A previous commit added requirement that changes to selected mpi slots
require a hard reset or restart to take effect.  This solved a
longstanding issue where the dialog allowed roms to be deselected
or removed while code was still executing in them which caused undefined
behaviour and sometimes memory access violations.

To make things easier for users a VCC reset button is added to the mpi
config.  Users can press the button for their changes to take immediate
effect.

Alternative idea was a "auto reset" checkbox that when checked permitted
the mpi to automatically reset when a change is made to a selected cart,
this was rejected because it was felt users might want to use the dialog
to make multiple changes before resetting.